requirements.txt
Flask==2.3.3
pymongo==4.5.0
qrcode==7.4.2
Pillow>=10.3.0
python-dotenv==1.0.0
dnspython==2.4.0
certifi==2024.2.2